services/model-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
mlflow==2.8.1
boto3==1.34.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
requests==2.31.0
httpx==0.25.2
sentencepiece==0.1.99
protobuf==4.25.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncio==3.4.3
joblib==1.3.2

# Tracing dependencies
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-redis==0.41b0
opentelemetry-instrumentation-psycopg2==0.41b0
psycopg2-binary==2.9.9
asyncpg==0.29.0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-jaeger-thrift==1.20.0
jaeger-client==4.8.0

# Explainability dependencies
shap==0.42.1
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.4

prometheus-client==0.19.0
